third_party/blink/web_tests/external/wpt/css/css-animations/AnimationEffect-updateTiming.tentative-expected.txt

This is a testharness.js-based test.
FAIL AnimationEffect.updateTiming({ duration }) causes changes to the animation-duration to be ignored assert_equals: Duration should be the value set by the API expected 2000 but got 4000
FAIL AnimationEffect.updateTiming({ iterations, direction }) causes changes to the animation-iteration-count and animation-direction to be ignored assert_equals: Iterations should be the value set by the API expected 2 but got 4
FAIL AnimationEffect.updateTiming({ delay, fill }) causes changes to the animation-delay and animation-fill-mode to be ignored assert_equals: Delay should be the value set by the API expected 2000 but got 4000
PASS AnimationEffect.updateTiming() does override to changes from animation-* properties if there is an error
FAIL AnimationEffect properties that do not map to animation-* properties should not be changed when animation-* style is updated assert_equals: endDelay should be the value set by the API expected "steps(4)" but got "linear"
Harness: the test ran to completion.
